storage

Module client_storage

Structs

ClientStorageThread 🔒

SqliteEngine 🔒

StorageShelf 🔒

https://storage.spec.whatwg.org/#storage-shelf

Constants

STORAGE_SHELF_QUOTA_BYTES 🔒

https://storage.spec.whatwg.org/#storage-quota The storage quota of a storage shelf is an implementation-defined conservative estimate of the total amount of byttes it can hold. We use 10 GiB per shelf, matching Firefox’s documented limit (https://developer.mozilla.org/en-US/docs/Web/API/Storage_API/Storage_quotas_and_eviction_criteria).

Traits

ClientStorageThreadFactory

RegistryEngine 🔒

Functions

bucket_mode 🔒

https://storage.spec.whatwg.org/#bucket-mode

create_a_storage_bucket 🔒

https://storage.spec.whatwg.org/#create-a-storage-bucket

create_a_storage_shelf 🔒

https://storage.spec.whatwg.org/#create-a-storage-shelf

directory_size 🔒

https://storage.spec.whatwg.org/#storage-usage

ensure_storage_shed 🔒

obtain_a_local_storage_shelf 🔒

https://storage.spec.whatwg.org/#obtain-a-local-storage-shelf

obtain_a_storage_shelf 🔒

https://storage.spec.whatwg.org/#obtain-a-storage-shelf

set_bucket_mode 🔒

https://storage.spec.whatwg.org/#dom-storagemanager-persist

storage_quota_for_bucket 🔒

https://storage.spec.whatwg.org/#storage-quota

storage_usage_for_bucket 🔒

https://storage.spec.whatwg.org/#storage-usage