Module storage::compaction::picker

Structs

LeveledTimeWindowPicker

LeveledTimeWindowPicker only handles level 0 to level 1 compaction in a time-window tiered manner. It picks all SSTs in level 0 and writes rows in these SSTs to a new file partitioned by a inferred time bucket in level 1.

PickerContext

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Traits

Picker

Picker picks input SST files and builds the compaction task. Different compaction strategy may implement different pickers.

Functions

calculate_time_buckets 🔒

Calculates buckets for files. If file does not contain a time range in metadata, it will be assigned to a special bucket i64::MAX (normally no timestamp can be aligned to this bucket) so that all files without timestamp can be compacted together.

file_time_bucket_span 🔒

Calculates timestamp span between start and end timestamp.

find_compactable_files 🔒

Finds files that can be compacted in given level. Currently they’re files that is not currently under compaction.

get_expired_ssts 🔒