storage::indexeddb::engines::sqlite

Module encoding

Enums

KeyType 🔒

Functions

decode_stringy 🔒

deserialize

deserialize_number

encode_stringy 🔒

When encoding strings, we use variable-size encoding per the following table

internal_deserialize 🔒

internal_serialize 🔒

Tracks type offsets for arrays to ensure that we don’t have collisions with type bytes. When we hit the limit, we push an extra byte to separate the levels.

serialize

serialize_number

When encoding floats, 64bit IEEE 754 are almost sortable, except that positive sort lower than negative, and negative sort descending. So we use the following encoding: