storage::indexeddb

Struct IndexedDBEnvironment

struct IndexedDBEnvironment<E: KvsEngine> {

    engine: E,
    manager_sender: GenericSender<IndexedDBThreadMsg>,
    transactions: FxHashMap<u64, KvsTransaction>,
    next_created_seq: u64,
    txn_info: FxHashMap<u64, TxnInfo>,
    queued_readwrite: VecDeque<u64>,
    queued_readonly: VecDeque<u64>,
    queued_readwrite_set: FxHashSet<u64>,
    queued_readonly_set: FxHashSet<u64>,
    running_readwrite: Option<u64>,
    running_readonly: HashSet<u64>,
    handled_next_unhandled_request_id: FxHashMap<u64, u64>,
    handled_pending: FxHashMap<u64, HashSet<u64>>,
    pending_commit_callbacks: FxHashMap<u64, Vec<GenericCallback<TxnCompleteMsg>>>,
    pending_abort_callbacks: FxHashMap<u64, Vec<GenericCallback<TxnCompleteMsg>>>,

}

Fields
engine: Emanager_sender: GenericSender<IndexedDBThreadMsg>transactions: FxHashMap<u64, KvsTransaction>next_created_seq: u64txn_info: FxHashMap<u64, TxnInfo>queued_readwrite: VecDeque<u64>queued_readonly: VecDeque<u64>queued_readwrite_set: FxHashSet<u64>queued_readonly_set: FxHashSet<u64>running_readwrite: Option<u64>running_readonly: HashSet<u64>handled_next_unhandled_request_id: FxHashMap<u64, u64>handled_pending: FxHashMap<u64, HashSet<u64>>pending_commit_callbacks: FxHashMap<u64, Vec<GenericCallback<TxnCompleteMsg>>>pending_abort_callbacks: FxHashMap<u64, Vec<GenericCallback<TxnCompleteMsg>>>
Implementations

impl<E: KvsEngine> IndexedDBEnvironment<E>

fn new( engine: E, manager_sender: GenericSender<IndexedDBThreadMsg>, ) -> IndexedDBEnvironment<E>

fn register_transaction( &mut self, txn: u64, mode: IndexedDBTxnMode, scope: Vec<String>, )

fn scopes_overlap(a: &TxnInfo, b: &TxnInfo) -> bool

fn earlier_overlapping_live_exists<F>( &self, transaction: u64, predicate: F, ) -> bool
where F: Fn(&TxnInfo) -> bool,

fn can_start_by_spec(&self, txn: u64) -> bool

fn enqueue_txn(&mut self, txn: u64, mode: &IndexedDBTxnMode)

fn queue_operation( &mut self, store_name: &str, serial_number: u64, mode: IndexedDBTxnMode, operation: AsyncOperation, )

fn schedule_transactions(&mut self, origin: ImmutableOrigin, db_name: &str)

https://w3c.github.io/IndexedDB/#transaction-scheduling

fn start_transaction( &mut self, origin: ImmutableOrigin, db_name: String, txn: u64, sender: Option<GenericSender<BackendResult<()>>>, )

fn mark_request_handled(&mut self, txn: u64, request_id: u64)

fn can_notify_txn_maybe_commit(&self, txn: u64) -> bool

fn can_commit_now(&self, txn: u64) -> bool

fn queue_pending_commit_callback( &mut self, txn: u64, callback: GenericCallback<TxnCompleteMsg>, )

fn take_pending_commit_callbacks( &mut self, txn: u64, ) -> Vec<GenericCallback<TxnCompleteMsg>>

fn maybe_commit_candidates(&self) -> Vec<u64>

fn finish_transaction(&mut self, txn: u64)

fn abort_transaction(&mut self, txn: u64)

fn key_generator_current_number(&self, store_name: &str) -> Option<i64>

fn set_key_generator_current_number( &self, store_name: &str, current_number: i64, ) -> DbResult<()>

fn restore_key_generators_after_abort( &self, key_generator_snapshots: &[KeyGeneratorSnapshot], ) -> DbResult<()>

https://w3c.github.io/IndexedDB/#key-generator-construct

fn object_store(&self, store_name: &str) -> DbResult<IndexedDBObjectStore>

https://w3c.github.io/IndexedDB/#key-generator-construct

fn object_stores(&self) -> DbResult<Vec<IndexedDBObjectStore>>

fn key_path(&self, store_name: &str) -> Option<KeyPath>

fn object_store_names(&self) -> DbResult<Vec<String>>

fn indexes(&self, store_name: &str) -> DbResult<Vec<IndexedDBIndex>>

fn create_index( &self, store_name: &str, index_name: String, key_path: KeyPath, unique: bool, multi_entry: bool, ) -> DbResult<CreateObjectResult>

fn delete_index(&self, store_name: &str, index_name: String) -> DbResult<()>

fn create_object_store( &mut self, store_name: &str, key_path: Option<KeyPath>, auto_increment: bool, ) -> DbResult<CreateObjectResult>

fn delete_object_store(&mut self, store_name: &str) -> DbResult<()>

fn restore_object_stores( &mut self, object_stores: &[IndexedDBObjectStore], ) -> DbResult<()>

Restores the object-store metadata snapshot captured before an aborted upgrade: remove stores/indexes created during the aborted upgrade, create stores/indexes that existed before it, and reset key-generator current numbers for stores that use key generators.

fn version(&self) -> Result<u64, E::Error>

fn set_version(&mut self, version: u64) -> DbResult<()>

fn abort( &mut self, origin: &ImmutableOrigin, database_name: &str, transaction: u64, ) -> bool

https://w3c.github.io/IndexedDB/#abort-a-transaction

When a transaction is aborted the implementation must undo (roll back) any changes that were made to the database during that transaction.

This only aborts the transaction if one was previously queued by adding an abort callback to Self::pending_abort_callbacks.

TODO: implement the abort algorithm and rollback for the engine.