storage::indexeddb

Struct IndexedDBManager

struct IndexedDBManager {
    port: GenericReceiver<IndexedDBThreadMsg>,
    manager_sender: GenericSender<IndexedDBThreadMsg>,
    databases: HashMap<IndexedDBDescription, IndexedDBEnvironment<SqliteEngine>>,
    thread_pool: Arc<ThreadPool>,
    serial_number_counter: u64,
    connection_queues: HashMap<IndexedDBDescription, VecDeque<OpenRequest>>,
    connections: HashMap<IndexedDBDescription, HashMap<Uuid, Connection>>,
}

Fields
port: GenericReceiver<IndexedDBThreadMsg>manager_sender: GenericSender<IndexedDBThreadMsg>databases: HashMap<IndexedDBDescription, IndexedDBEnvironment<SqliteEngine>>thread_pool: Arc<ThreadPool>serial_number_counter: u64

A global counter to produce unique transaction ids. TODO: remove once db connections lifecyle is managed. A global counter is only necessary because of how deleting a db currently does not wait for connection to close and transactions to finish.

connection_queues: HashMap<IndexedDBDescription, VecDeque<OpenRequest>>

https://w3c.github.io/IndexedDB/#connection-queue

connections: HashMap<IndexedDBDescription, HashMap<Uuid, Connection>>

https://w3c.github.io/IndexedDB/#connection

Implementations

impl IndexedDBManager

fn new( port: GenericReceiver<IndexedDBThreadMsg>, manager_sender: GenericSender<IndexedDBThreadMsg>, ) -> IndexedDBManager

impl IndexedDBManager

fn start(&mut self)

fn dispatch_txn_maybe_commit( &self, origin: ImmutableOrigin, db_name: String, txn: u64, )

fn handle_txn_maybe_commit( &mut self, origin: ImmutableOrigin, db_name: String, txn: u64, )

fn handle_upgrade_transaction_finished( &mut self, name: String, origin: ImmutableOrigin, txn: u64, committed: bool, )

Handle when an upgrade transaction finishes in script.

fn advance_connection_queue(&mut self, key: IndexedDBDescription)

Run the next open request in the queue.

fn maybe_remove_front_from_queue(&mut self, key: &IndexedDBDescription) -> bool

Remove the record at the front if it is not pending.

fn remove_connection(&mut self, key: &IndexedDBDescription, id: &Uuid)

fn revert_aborted_upgrade( &mut self, key: &IndexedDBDescription, upgrade: &VersionUpgrade, proxy_map: &StorageProxyMap, )

Revert the backing database state after aborting an upgrade transaction.

https://w3c.github.io/IndexedDB/#abort-an-upgrade-transaction IndexedDB 5.8 step 3 restores the previous version, or 0 if the database was newly created. Step 4 restores the previous object store set, or the empty set if the database was newly created. Servo eagerly creates the backing database with version 0 and no stores during open, so aborting that first upgrade must roll back to the pre-creation state by deleting the placeholder backing store entirely.

Related: https://github.com/servo/servo/pull/42998

fn abort_pending_upgrade( &mut self, name: String, id: Uuid, origin: ImmutableOrigin, proxy_map: &StorageProxyMap, )

Aborting the current upgrade for an origin.

fn abort_pending_upgrades( &mut self, pending_upgrades: HashMap<String, HashSet<Uuid>>, origin: ImmutableOrigin, proxy_map: StorageProxyMap, )

Aborting all upgrades for an origin Note: this only reverts the version at this point.

fn open_a_database_connection( &mut self, sender: GenericCallback<ConnectionMsg>, origin: ImmutableOrigin, db_name: String, version: Option<u64>, id: Uuid, proxy_map: StorageProxyMap, )

https://w3c.github.io/IndexedDB/#open-a-database-connection

fn get_database( &self, origin: ImmutableOrigin, db_name: String, ) -> Option<&IndexedDBEnvironment<SqliteEngine>>

fn get_database_mut( &mut self, origin: ImmutableOrigin, db_name: String, ) -> Option<&mut IndexedDBEnvironment<SqliteEngine>>

fn upgrade_database(&mut self, key: IndexedDBDescription, new_version: u64)

https://w3c.github.io/IndexedDB/#upgrade-a-database To upgrade a database with connection (a connection), a new version, and a request, run these steps:

fn handle_version_change_done( &mut self, name: String, from_id: Uuid, old_version: u64, origin: ImmutableOrigin, )

https://w3c.github.io/IndexedDB/#open-a-database-connection

fn open_database(&mut self, key: IndexedDBDescription)

https://w3c.github.io/IndexedDB/#open-a-database-connection The part where the open request is ready for processing.

fn start_delete_database( &mut self, key: IndexedDBDescription, id: Uuid, proxy_map: StorageProxyMap, sender: GenericCallback<BackendResult<u64>>, )

https://www.w3.org/TR/IndexedDB/#delete-a-database The part adding the request to the connection queue.

fn delete_database(&mut self, key: IndexedDBDescription)

https://www.w3.org/TR/IndexedDB/#delete-a-database

fn close_database(&mut self, origin: ImmutableOrigin, id: Uuid, name: String)

https://w3c.github.io/IndexedDB/#closing-connection

fn handle_sync_operation(&mut self, operation: SyncOperation)

fn handle_abort( &mut self, abort_callback: GenericCallback<TxnCompleteMsg>, origin: ImmutableOrigin, database_name: String, transaction: u64, )

Handling for the Abort message which will call [Self::abort] if the transaction being aborted is not ongoing. If the transaction is in process, abort is delayed until the batch finishes.

fn collect_memory_reports(&self) -> Vec<Report>