Trait storage::memtable::Memtable

pub trait Memtable: Send + Sync + Debug {
    // Required methods
    fn id(&self) -> MemtableId;
    fn schema(&self) -> RegionSchemaRef;
    fn write(&self, kvs: &KeyValues) -> Result<()>;
    fn iter(&self, ctx: &IterContext) -> Result<BoxedBatchIterator>;
    fn bytes_allocated(&self) -> usize;
    fn num_rows(&self) -> usize;
}

In memory storage.

Required Methods

fn id(&self) -> MemtableId

Returns id of this memtable.

fn schema(&self) -> RegionSchemaRef

Returns schema of the memtable.

fn write(&self, kvs: &KeyValues) -> Result<()>

Write key/values to the memtable.

Panics

Panics if the schema of key/value differs from memtable’s schema.

fn iter(&self, ctx: &IterContext) -> Result<BoxedBatchIterator>

Iterates the memtable.

fn bytes_allocated(&self) -> usize

Returns the estimated bytes allocated by this memtable from heap. Result of this method may be larger than the estimated based on [num_rows] because of the implementor’s pre-alloc behavior.

fn num_rows(&self) -> usize

Return the number of rows contained in this memtable.