Module storage::read

Common structs and utilities for read.

Modules

chain 🔒

dedup 🔒

merge 🔒

Merge reader.

windowed 🔒

Structs

Batch

Storage internal representation of a batch of rows.

BatchBuilder

Reusable Batch builder.

ChainReader

A reader that simply chain the outputs of input readers.

DedupReader

A reader that dedup rows from inner reader.

MergeReader

A reader that would sort and merge Batch from multiple sources by key.

MergeReaderBuilder

WindowedReader

WindowedReader provides a windowed record batch reader that scans all rows within a window at a time and sort these rows ordered in [<timestamp>, <PK>] order.

Traits

BatchOp

Compute operations for Batch.

BatchReader

Async batch reader.

Type Definitions

BoxedBatchReader

Pointer to BatchReader.