Trait storage::read::BatchOp

pub trait BatchOp {
    // Required methods
    fn compare_row(
        &self,
        left: &Batch,
        i: usize,
        right: &Batch,
        j: usize
    ) -> Ordering;
    fn find_unique(
        &self,
        batch: &Batch,
        selected: &mut BitVec,
        prev: Option<&Batch>
    );
    fn filter(&self, batch: &Batch, filter: &BooleanVector) -> Result<Batch>;
    fn unselect_deleted(&self, batch: &Batch, selected: &mut BitVec);
}

Compute operations for Batch.

Required Methods

fn compare_row(&self, left: &Batch, i: usize, right: &Batch, j: usize) -> Ordering

Compare i-th in left to j-th row in right by key (row key + internal columns).

The caller should ensure left and right have same schema as self.

Panics

Panics if

i or j is out of bound.

left or right has insufficient column num.

fn find_unique(&self, batch: &Batch, selected: &mut BitVec, prev: Option<&Batch>)

Find unique rows in batch by row key.

If prev is Some and not empty, the last row of prev would be used to dedup current batch. Set i-th bit of selected to true if i-th row is unique, which means the row key of i-th row is different from i+1-th’s.

The caller could use selected to build a BooleanVector to filter the batch, and must ensure selected is initialized by filling batch.num_rows() bits to zero.

Panics

Panics if

batch and prev have different number of columns (unless prev is empty).

selected.len() is less than the number of rows.

fn filter(&self, batch: &Batch, filter: &BooleanVector) -> Result<Batch>

Filters the batch, returns elements matching the filter (i.e. where the values are true).

Note that the nulls of filter are interpreted as false will lead to these elements being masked out.

fn unselect_deleted(&self, batch: &Batch, selected: &mut BitVec)

Unselect deleted rows according to the OpType.

Panics

Panics if

batch doesn’t have a valid op type column.

selected.len() is less than the number of rows.