Trait storage::read::BatchReader

pub trait BatchReader: Send {
    // Required method
    fn next_batch<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Batch>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Async batch reader.

Required Methods

fn next_batch<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<Option<Batch>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Fetch next Batch.

Returns Ok(None) when the reader has reached its end and calling next_batch() again won’t return batch again.

If Err is returned, caller should not call this method again, the implementor may or may not panic in such case.