Trait storage::schema::compat::CompatWrite

pub trait CompatWrite {
    // Required method
    fn compat_write(&mut self, dest_schema: &SchemaRef) -> Result<()>;
}

Make schema compatible to write to target with another schema.

Required Methods

fn compat_write(&mut self, dest_schema: &SchemaRef) -> Result<()>

Makes the schema of self compatible with dest_schema.

For column in dest_schema but not in self, this method would insert a vector with default value.

If there are columns not in dest_schema, an error would be returned.