Module storage::sst

Modules

parquet 🔒

Parquet sst format.

pruning 🔒

stream_writer 🔒

Structs

BufferedWriter

Parquet writer that buffers row groups in memory and writes buffered data to an underlying storage by chunks to reduce memory consumption.

FileHandle

FileHandleInner 🔒

Actually data of FileHandle.

FileId

Unique id for [SST File].

FileMeta

Immutable metadata of a sst file.

FsAccessLayer

Sst access layer.

LazyParquetBatchReader 🔒

LevelMeta

Metadata of files in same SST level.

LevelMetas

Metadata of all SSTs under a region.

ParseIdError

ParseIdSnafu 🔒

SNAFU context selector for the ParseIdError error

ReadOptions

SstInfo

WriteOptions

Enums

Source

Parquet writer data source.

Constants

MAX_LEVEL

Maximum level of SSTs.

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

Traits

AccessLayer

SST access layer.

Functions

deserialize_from_string 🔒

new_level_meta_vec 🔒

Type Definitions

AccessLayerRef

Level

LevelMetaVec 🔒