Struct storage::storage::Storage

pub struct Storage {
    root_storage_dir: String,
}

Storage struct

Fields
root_storage_dir: String

the root directory of the storage server

Implementations

impl Storage

Storage implementation

pub fn new(root: &Path) -> Self

pub async fn initialize_storage( &self, client_port: u16, command_port: u16, registration_port: u16, root_storage_dir: &Path ) -> Result<(), Box<dyn Error>>

Initialize the storage server by registering with the naming server and deleting files that are not in the naming server

pub fn remove_dir_recursive(&self, path: &Path) -> Result<(), Error>

Remove a directory if it is empty and recursively remove its parent directories if they are empty

pub fn is_dir_empty(&self, path: &Path) -> Result<bool, Error>

Check if a directory is empty

pub fn get_all_files_recursive( &self, path: &Path ) -> Result<Vec<PathBuf>, Error>

Get all files in the storage directory recursively

fn get_full_path( &self, path: &Path ) -> Result<PathBuf, (StatusCode, Json<ExceptionReturn>)>

Get the full path of a file with the given path

pub fn get_file_size( &self, path: &Path ) -> Result<u64, (StatusCode, Json<ExceptionReturn>)>

Get the size of a file with the given path

pub fn find_file( &self, path: &Path ) -> Result<File, (StatusCode, Json<ExceptionReturn>)>

Find a file with the given path

pub fn read( &self, path: &Path, offset: u64, length: u64 ) -> Result<Vec<u8>, (StatusCode, Json<ExceptionReturn>)>

Read data from a file with the given path, offset, and length

pub fn write( &self, path: &Path, offset: i64, data: Vec<u8> ) -> Result<(), (StatusCode, Json<ExceptionReturn>)>

Write data to a file with the given path, offset, and data

pub fn create_file( &self, path: &Path ) -> Result<bool, (StatusCode, Json<ExceptionReturn>)>

Create a file with the given path

pub fn delete_file( &self, path: &Path ) -> Result<bool, (StatusCode, Json<ExceptionReturn>)>

Delete a file with the given path

pub async fn copy( &self, path: &Path, server_ip: &str, server_port: u16 ) -> Result<bool, (StatusCode, Json<ExceptionReturn>)>

Copy a file from another storage server with the given path, server IP, and server port

pub fn is_valid_path(path: &Path) -> bool

Check if a path is valid

pub fn create_file_ignore_exist( &self, path: &Path ) -> Result<File, (StatusCode, Json<ExceptionReturn>)>

Create a file with the given path if it does not exist, if it exists, delete it and create a new one