Struct storage::CertificateStore

pub struct CertificateStore { /* private fields */ }

The main storage when we have to deal with certificates. It maintains two storages, one main which saves the certificates by their ids, and a secondary one which acts as an index to allow us fast retrieval based for queries based in certificate rounds. It also offers pub/sub capabilities in write events. By using the notify_read someone can wait to hear until a certificate by a specific id has been written in storage.

Implementations

impl CertificateStore

pub fn new(
certificates_by_id: DBMap<CertificateDigest, Certificate>,
certificate_ids_by_round: DBMap<(Round, CertificateDigest), CertificateToken>
) -> CertificateStore

pub fn write(&self, certificate: Certificate) -> StoreResult<()>

Inserts a certificate to the store

pub fn write_all(
&self,
certificates: impl IntoIterator<Item = Certificate>
) -> StoreResult<()>

Inserts multiple certificates in the storage. This is an atomic operation. In the end it notifies any subscribers that are waiting to hear for the value.

pub fn read(&self, id: CertificateDigest) -> StoreResult<Option<Certificate>>

Retrieves a certificate from the store. If not found then None is returned as result.

pub fn read_all(
&self,
ids: impl IntoIterator<Item = CertificateDigest>
) -> StoreResult<Vec<Option<Certificate>>>

Retrieves multiple certificates by their provided ids. The results are returned in the same sequence as the provided keys.

pub async fn notify_read(
&self,
id: CertificateDigest
) -> StoreResult<Certificate>

Waits to get notified until the requested certificate becomes available

pub fn delete(&self, id: CertificateDigest) -> StoreResult<()>

Deletes a single certificate by its digest.

pub fn delete_all(
&self,
ids: impl IntoIterator<Item = CertificateDigest>
) -> StoreResult<()>

Deletes multiple certificates in an atomic way.

pub fn after_round(&self, round: Round) -> StoreResult<Vec<Certificate>>

Retrieves all the certificates with round >= the provided round. The result is returned with certificates sorted in round asc order

pub fn last_round(&self) -> StoreResult<Vec<Certificate>>

Retrieves the certificates of the last round

pub fn clear(&self) -> StoreResult<()>

Clears both the main storage of the certificates and the secondary index

pub fn is_empty(&self) -> bool

Checks whether the storage is empty. The main storage is being used to determine this.