storage

Struct DiskFileManager

pub struct DiskFileManager {
    files: RwLock<HashMap<u32, Arc<File>>>,
    file_catalog: Arc<FileCatalog>,
}

DiskFileManager struct, an implementation of FileManager A disk based file manager

Fields
files: RwLock<HashMap<u32, Arc<File>>>file_catalog: Arc<FileCatalog>
Implementations

impl DiskFileManager

fn get_or_open_file(&self, file_id: u32) -> Result<Arc<File>, FileManagerError>

Attempt to either retrieve the file handle or open one if not already present in memory Returns Arc<File> if successful, FileManagerError if something goes wrong

fn read_at(file: &File, buf: &mut [u8], offset: u64) -> Result<usize>

OS Specific method to read an array of bytes from a provided offset in the file Returns the number of bytes read

fn write_at(file: &File, buf: &[u8], offset: u64) -> Result<usize>

OS Specific method to write a byte array at a provided offset in the file Returns the number of bytes written

fn ensure_parent_dir(path: &Path)

Ensures the data directory storing all the database files exists If it does not exist, it creates it.

Returns

() - nothing, as the method will cause a panic if the directory cannot be created.