Struct storage::StorageHandler

pub struct StorageHandler {
    directory: PathBuf,
    pool: SqlitePool,
    size_threshold: f64,
    cleanup_target: f64,
}

Fields
directory: PathBuf

Root directory to watch

Database will be stored in here

pool: SqlitePoolsize_threshold: f64

Directory size limit in bytes above which a call to .maybe_cleanup() triggers

cleanup_target: f64

Target amount of bytes to delete during cleanup

Implementations

impl StorageHandler

pub async fn storage_id(directory: PathBuf) -> Result<String, StorageError>

Fetches the storage ID of the given directory. If the directory has not previously been used as a storage a new identifier will be created and written to disk.

pub async fn new(
directory: PathBuf,
size_threshold: f64,
cleanup_target: f64
) -> Result<Self, StorageError>

Creates new instance that watches a given directory and attempts to clean at least cleanup_target bytes if the watched directory grows larger than size_threshold bytes

pub async fn scan_fs<'_>(&'_ self) -> Result<(), StorageError>

Explicitly scan the full filesystem and sync the database

This should only be called infrequently as it usually is an expensive operation.

pub async fn add_file<'_, P: AsRef<Path>>(
&'_ self,
path: P
) -> Result<(), StorageError>

Import or update a file to the database

Reads the files metadata from disk and upserts the value into the database.

pub async fn remove_file<'_, P: AsRef<Path>>(
&'_ self,
path: P
) -> Result<(), StorageError>

Remove a file from the database

Call this if you externally delete a file from the watched directory

pub async fn maybe_cleanup<'_>(&'_ self) -> Result<usize, StorageError>

Runs a cleanup if the used bytes exceed the size_threshold

async fn cleanup<'_>(&'_ self, used_bytes: f64) -> Result<usize, StorageError>

async fn remove_file_from_con<'_, P: AsRef<Path>, E: Executor<Database = Sqlite>>(
&'_ self,
path: P,
__arg2: E
) -> Result<(), StorageError>

fn relative_path<P: AsRef<Path>>(
&self,
path: P
) -> Result<PathBuf, StorageError>

async fn acquire_connection<'_>(
&'_ self
) -> Result<PoolConnection<SqliteConnection>, StorageError>