storage::webstorage

Struct WebStorageManager

struct WebStorageManager {
    port: GenericReceiver<WebStorageThreadMsg>,
    session_storage_origins: StorageOrigins,
    local_storage_origins: StorageOrigins,
    session_data: FxHashMap<WebViewId, FxHashMap<ImmutableOrigin, OriginEntry>>,
    config_dir: Option<PathBuf>,
    thread_pool: Arc<ThreadPool>,
    environments: FxHashMap<ImmutableOrigin, WebStorageEnvironment<SqliteEngine>>,
}

Fields
port: GenericReceiver<WebStorageThreadMsg>session_storage_origins: StorageOriginslocal_storage_origins: StorageOriginssession_data: FxHashMap<WebViewId, FxHashMap<ImmutableOrigin, OriginEntry>>config_dir: Option<PathBuf>thread_pool: Arc<ThreadPool>environments: FxHashMap<ImmutableOrigin, WebStorageEnvironment<SqliteEngine>>
Implementations

impl WebStorageManager

fn new( port: GenericReceiver<WebStorageThreadMsg>, config_dir: Option<PathBuf>, ) -> WebStorageManager

impl WebStorageManager

fn start(&mut self)

fn collect_memory_reports(&self) -> Vec<Report>

fn save_local_storage_origins(&self)

fn get_origin_location(&self, origin: &ImmutableOrigin) -> Option<PathBuf>

fn add_new_environment(&mut self, origin: &ImmutableOrigin) -> Result<(), Error>

fn get_environment( &mut self, origin: &ImmutableOrigin, ) -> Result<&WebStorageEnvironment<SqliteEngine>, Error>

fn get_environment_mut( &mut self, origin: &ImmutableOrigin, ) -> Result<&mut WebStorageEnvironment<SqliteEngine>, Error>

fn select_data( &mut self, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, ) -> Option<&OriginEntry>

fn select_data_mut( &mut self, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, ) -> Option<&mut OriginEntry>

fn ensure_data_mut( &mut self, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, ) -> Option<&mut OriginEntry>

fn length( &mut self, sender: GenericSender<usize>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, )

fn key( &mut self, sender: GenericSender<Option<String>>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, index: u32, )

fn keys( &mut self, sender: GenericSender<Vec<String>>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, )

fn set_item( &mut self, sender: GenericSender<Result<(bool, Option<String>), ()>>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, name: String, value: String, )

Sends Ok(changed, Some(old_value)) in case there was a previous value with the same key name but with different value name otherwise sends Err(()) to indicate that the operation would result in exceeding the quota limit

fn request_item( &mut self, sender: GenericSender<Option<String>>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, name: String, )

fn remove_item( &mut self, sender: GenericSender<Option<String>>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, name: String, )

Sends Some(old_value) in case there was a previous value with the key name, otherwise sends None

fn clear( &mut self, sender: GenericSender<bool>, storage_type: WebStorageType, webview_id: WebViewId, origin: ImmutableOrigin, )

fn clone(&mut self, src_webview_id: WebViewId, dest_webview_id: WebViewId)

fn origin_descriptors( &mut self, storage_type: WebStorageType, ) -> Vec<OriginDescriptor>

fn clear_data_for_sites( &mut self, storage_type: WebStorageType, sites: &[String], )