Module storage::write_batch

Modules

codec

compat 🔒

Structs

Mutation

A write operation to the region.

NameToVector 🔒

Vectors in NameToVector have same length.

Payload

Data of WriteBatch.

WriteBatch

Implementation of [WriteRequest].

Constants

MAX_BATCH_SIZE 🔒

Max number of updates in a write batch.

Functions

first_vector_len 🔒

Returns the length of the first vector in data.

new_column_with_default_value 🔒

Creates a new column and fills it by default value.

validate_column 🔒

Checks whether col matches given column_schema.